+++ src/chains/Base/Common.sol
@@ -11,7 +11,12 @@ import {IPoolManager} from "../../core/UniswapV4Types.sol";
 import {EulerSwap, IEVC, IEulerSwap} from "../../core/EulerSwap.sol";
 import {BalancerV3} from "../../core/BalancerV3.sol";
 import {PancakeInfinity} from "../../core/PancakeInfinity.sol";
-import {Renegade, BASE_SELECTOR} from "../../core/Renegade.sol";
+import {
+    pancakeInfinityVault,
+    pancakeInfinityClManager,
+    pancakeInfinityBinManager
+} from "../../core/pancakeInfinityForks/PancakeInfinity.sol";
+import {Renegade} from "../../core/Renegade.sol";
 import {Bebop} from "../../core/Bebop.sol";
 import {Hanji} from "../../core/Hanji.sol";
 
@@ -76,6 +81,10 @@ abstract contract BaseMixin is
 
     constructor() {
         assert(block.chainid == 8453 || block.chainid == 31337);
+    }
+
+    function _renegadeGasSponsorV2() internal pure override returns (address) {
+        return 0xD9E0507D706408D0f14E22e50880189Fd915be80;
     }
 
     function _dispatch(uint256 i, uint256 action, bytes calldata data, AllowedSlippage memory slippage)
@@ -145,10 +154,27 @@ abstract contract BaseMixin is
 
             sellToDodoV2(recipient, sellToken, ppm, dodo, quoteForBase, minBuyAmount);
         } else if (action == uint32(ISettlerActions.RENEGADE.selector)) {
-            (address target, IERC20 sellToken, bool baseForQuote, bytes memory renegadeData, uint256 minBuyAmount) =
-                abi.decode(data, (address, IERC20, bool, bytes, uint256));
+            (
+                address recipient,
+                IERC20 sellToken,
+                IERC20 buyToken,
+                uint256 maxSellAmount,
+                bool refundNativeEth,
+                uint256 maxRefundAmount,
+                bytes memory renegadeData,
+                uint256 minBuyAmount
+            ) = abi.decode(data, (address, IERC20, IERC20, uint256, bool, uint256, bytes, uint256));
 
-            sellToRenegade(target, sellToken, baseForQuote, renegadeData, minBuyAmount);
+            sellToRenegade(
+                recipient,
+                sellToken,
+                buyToken,
+                maxSellAmount,
+                refundNativeEth,
+                maxRefundAmount,
+                renegadeData,
+                minBuyAmount
+            );
         } else if (action == uint32(ISettlerActions.HANJI.selector)) {
             (
                 IERC20 sellToken,
@@ -227,6 +253,18 @@ abstract contract BaseMixin is
         return BASE_POOL_MANAGER;
     }
 
+    function _PANCAKE_INFINITY_VAULT() internal pure override returns (address) {
+        return pancakeInfinityVault;
+    }
+
+    function _PANCAKE_INFINITY_CL_MANAGER() internal pure override returns (address) {
+        return pancakeInfinityClManager;
+    }
+
+    function _PANCAKE_INFINITY_BIN_MANAGER() internal pure override returns (address) {
+        return pancakeInfinityBinManager;
+    }
+
     /*
     function _EVC() internal pure override returns (IEVC) {
         return IEVC(0x5301c7dD20bD945D2013b48ed0DEE3A284ca8989);
@@ -255,10 +293,6 @@ abstract contract BaseMixin is
             mstore(returndata, 0x20)
             mstore(add(0x20, returndata), shr(0x60, msgSenderShifted))
         }
-    }
-
-    function _renegadeSelector() internal pure override returns (uint32) {
-        return BASE_SELECTOR;
     }
 
     // I hate Solidity inheritance